+++ stock/src/main/java/app/models/Item.java
@@ -10,11 +10,10 @@ public class Item {
     @PartitionKey
     public UUID item_id = UUID.randomUUID();
 
-
     public int stock;
 
-    public int price;
-    public Item(String itemId, int stock, int price) {
+    public float price;
+    public Item(String itemId, int stock, float price) {
         this.item_id = UUID.fromString(itemId);
         this.stock = stock;
         this.price = price;
